feat: Modal CT
